Add automatic updating to Komet
Updates are installed after the application quits, unless updates can't be installed without authentication. If authentication is required, updates just aren't installed automatically, but the user can manually check for updates if they wish to.

I wrote custom user interface for the installer using my fork of Sparkle. This allows the user (if they initiate an update) to have an update be installed on quit.

We want to at all costs not be intrusive.