Fix janky scrollbar position
If scrollbars are set to 'Always' under system preferences they were
appearing with a weird offset to the top and right. I'd cheated and used
trailing+top space around the scroll view to offset text; this commit
changes those offsets to be what they ought to be, and instead uses
NSTextView's own insets.