Fix view jittery bug by using temp attributes for overflow color
When some kind of (potentially long?) content is loaded, the view could jitter around when typing in text. This was caused by some combination of disabling drawsInBackground and removing the background color attribute. We workaround this by removing/adding temp attributes for text overflow changes (if a line exceeds a certain number of characters).