Add a preference to enable/disable autocorrect.
This preference defaults to the system's automatic spelling correction
setting; toggling the checkbox in Komet's preferences window will
override this setting to selectively enable or disable automatic
spelling correction in Komet.

This commit also removes the existing default key for autocorrection
(ZGCommitTextViewAutomaticSpellingCorrection) and replaces it with a
ZGUserDefaults key (ZGEditorEnableAutomaticSpellingCorrection) to
reflect the fact that this default is now user-configurable.